# openapi/pvlog-v1.yaml
openapi: 3.1.1
...
info:
  title: PVLog REST API
  version: 0.1.0
  summary: Modern self-hosted photovoltaic monitoring API
  description: >-
    The versioned JSON contract for PVLog. PVOutput r2 compatibility endpoints
    are documented separately because they use legacy form and CSV formats.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
servers:
  - url: /
    description: Current PVLog installation
...
paths:
...
  /api/v1/systems:
    post:
      operationId: createSystem
      tags: [Systems]
      summary: Create a private PV system
      description: Creates one account-owned system using private and active defaults.
      security: [{ bearerAuth: [systems:write] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSystem" }
            example: { accountId: 019505c8-7c85-7f0b-9bc3-2a3c4d5e6f70, name: Roof PV, timezone: Europe/Berlin }
      responses:
        "201": { $ref: "#/components/responses/SystemResponse" }
        "422": { $ref: "#/components/responses/ValidationProblem" }
  /api/v1/systems/{id}:
    parameters:
      - $ref: "#/components/parameters/SystemId"
    put:
      operationId: updateSystem
      tags: [Systems]
      summary: Update a PV system
      description: Replaces mutable system settings when the supplied ETag is current.
      security: [{ bearerAuth: [systems:write] }]
      parameters: [{ $ref: "#/components/parameters/IfMatch" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSystem" }
      responses:
        "200": { $ref: "#/components/responses/SystemResponse" }
        "412": { $ref: "#/components/responses/PreconditionProblem" }
    delete:
      operationId: deleteSystem
      tags: [Systems]
      summary: Permanently delete a PV system
      description: Starts confirmed permanent deletion when the supplied ETag is current.
      security: [{ bearerAuth: [systems:write] }]
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - { name: X-Confirm-Delete, in: header, required: true, schema: { type: boolean } }
      responses:
        "204": { description: System deletion accepted }
        "412": { $ref: "#/components/responses/PreconditionProblem" }
  /api/v1/systems/{id}/archive:
    post:
      operationId: archiveSystem
      tags: [Systems]
      summary: Archive a PV system
      description: Makes an active system inactive without deleting its stored data.
      security: [{ bearerAuth: [systems:write] }]
      parameters:
        - $ref: "#/components/parameters/SystemId"
        - $ref: "#/components/parameters/IfMatch"
      responses: { "200": { $ref: "#/components/responses/SystemResponse" }, "412": { $ref: "#/components/responses/PreconditionProblem" } }
  /api/v1/systems/{id}/restore:
    post:
      operationId: restoreSystem
      tags: [Systems]
      summary: Restore a PV system
      description: Returns an archived system to active service.
      security: [{ bearerAuth: [systems:write] }]
      parameters:
        - $ref: "#/components/parameters/SystemId"
        - $ref: "#/components/parameters/IfMatch"
      responses: { "200": { $ref: "#/components/responses/SystemResponse" }, "412": { $ref: "#/components/responses/PreconditionProblem" } }
  /api/v1/accounts/{account_id}/systems/{system_id}/equipment:
    parameters: [{ $ref: "#/components/parameters/AccountId" }, { $ref: "#/components/parameters/NestedSystemId" }]
    get: { operationId: listEquipment, summary: List equipment, description: Lists effective-dated equipment for one system., tags: [Systems], security: [{ bearerAuth: [systems:read] }], responses: { "200": { $ref: "#/components/responses/ResourceList" }, "403": { $ref: "#/components/responses/ForbiddenProblem" } } }
    post: { operationId: createEquipment, summary: Create equipment, description: Adds validated effective-dated equipment to one system., tags: [Systems], security: [{ bearerAuth: [systems:write] }], requestBody: { $ref: "#/components/requestBodies/ManagedResource" }, responses: { "201": { $ref: "#/components/responses/ManagedResource" }, "403": { $ref: "#/components/responses/ForbiddenProblem" } } }
  /api/v1/accounts/{account_id}/systems/{system_id}/tariffs:
    parameters: [{ $ref: "#/components/parameters/AccountId" }, { $ref: "#/components/parameters/NestedSystemId" }]
    get: { operationId: listTariffs, summary: List tariffs, description: Lists effective-dated tariffs for one system., tags: [Systems], security: [{ bearerAuth: [systems:read] }], responses: { "200": { $ref: "#/components/responses/ResourceList" }, "403": { $ref: "#/components/responses/ForbiddenProblem" } } }
    post: { operationId: createTariff, summary: Create tariff, description: Adds a validated effective-dated tariff to one system., tags: [Systems], security: [{ bearerAuth: [systems:write] }], requestBody: { $ref: "#/components/requestBodies/ManagedResource" }, responses: { "201": { $ref: "#/components/responses/ManagedResource" }, "403": { $ref: "#/components/responses/ForbiddenProblem" } } }
  /api/v1/accounts/{account_id}/systems/{system_id}/channels:
    parameters: [{ $ref: "#/components/parameters/AccountId" }, { $ref: "#/components/parameters/NestedSystemId" }]
    get: { operationId: listChannels, summary: List channels, description: Lists typed extended channels for one system., tags: [Systems], security: [{ bearerAuth: [systems:read] }], responses: { "200": { $ref: "#/components/responses/ResourceList" }, "403": { $ref: "#/components/responses/ForbiddenProblem" } } }
    post: { operationId: createChannel, summary: Create channel, description: Adds a validated typed extended channel to one system., tags: [Systems], security: [{ bearerAuth: [systems:write] }], requestBody: { $ref: "#/components/requestBodies/ManagedResource" }, responses: { "201": { $ref: "#/components/responses/ManagedResource" }, "403": { $ref: "#/components/responses/ForbiddenProblem" } } }
  /api/v1/accounts/{account_id}/memberships:
    parameters: [{ $ref: "#/components/parameters/AccountId" }]
    get: { operationId: listMemberships, summary: List memberships, description: Lists authorized memberships in one account., tags: [Systems], security: [{ bearerAuth: [systems:read] }], responses: { "200": { $ref: "#/components/responses/ResourceList" }, "403": { $ref: "#/components/responses/ForbiddenProblem" } } }
    post: { operationId: createMembership, summary: Create membership, description: Adds a validated membership to one account., tags: [Systems], security: [{ bearerAuth: [systems:write] }], requestBody: { $ref: "#/components/requestBodies/ManagedResource" }, responses: { "201": { $ref: "#/components/responses/ManagedResource" }, "403": { $ref: "#/components/responses/ForbiddenProblem" } } }
  /api/v1/accounts/{account_id}/credentials:
    parameters: [{ $ref: "#/components/parameters/AccountId" }]
    get: { operationId: listCredentials, summary: List credentials, description: Lists non-secret credential metadata in one account., tags: [Systems], security: [{ bearerAuth: [systems:read] }], responses: { "200": { $ref: "#/components/responses/ResourceList" }, "403": { $ref: "#/components/responses/ForbiddenProblem" } } }
    post: { operationId: createCredential, summary: Create credential, description: Creates a scoped credential and returns its secret only once., tags: [Systems], security: [{ bearerAuth: [systems:write] }], requestBody: { $ref: "#/components/requestBodies/ManagedResource" }, responses: { "201": { $ref: "#/components/responses/ManagedResource" }, "403": { $ref: "#/components/responses/ForbiddenProblem" } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    SystemId: { name: id, in: path, required: true, schema: { $ref: "#/components/schemas/UuidV7" } }
    AccountId: { name: account_id, in: path, required: true, schema: { $ref: "#/components/schemas/UuidV7" } }
    NestedSystemId: { name: system_id, in: path, required: true, schema: { $ref: "#/components/schemas/UuidV7" } }
...
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string }, example: '"3"' }
  requestBodies:
    ManagedResource:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
          example: { name: Example resource }
  responses:
    SystemResponse:
      description: Versioned system representation
      headers: { ETag: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: "#/components/schemas/System" } } }
    ManagedResource:
      description: Versioned managed resource
      headers: { ETag: { schema: { type: string } } }
      content: { application/json: { schema: { $ref: "#/components/schemas/ManagedResource" } } }
    ResourceList:
      description: Managed resource collection
      content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/ManagedResource" } } } }
    ValidationProblem: { description: Request validation failed, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    PreconditionProblem: { description: ETag precondition failed, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    ForbiddenProblem: { description: Required scope is missing, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
...
  schemas:
    CreateSystem:
      type: object
      required: [accountId, name, timezone]
      properties: { accountId: { $ref: "#/components/schemas/UuidV7" }, name: { type: string, minLength: 1 }, timezone: { type: string, example: Europe/Berlin } }
    UpdateSystem:
      type: object
      required: [name, timezone, visibility]
      properties: { name: { type: string, minLength: 1 }, timezone: { type: string }, visibility: { enum: [private, account, unlisted, public] } }
    System:
      allOf:
        - $ref: "#/components/schemas/UpdateSystem"
        - type: object
          required: [id, accountId, lifecycle, version]
          properties: { id: { $ref: "#/components/schemas/UuidV7" }, accountId: { $ref: "#/components/schemas/UuidV7" }, lifecycle: { enum: [active, archived, pending_deletion] }, version: { type: integer, minimum: 1 } }
    ManagedResource:
      type: object
      required: [id, accountId, kind, version, attributes]
      properties: { id: { $ref: "#/components/schemas/UuidV7" }, accountId: { $ref: "#/components/schemas/UuidV7" }, systemId: { $ref: "#/components/schemas/UuidV7" }, kind: { enum: [equipment, tariff, channel, membership, credential] }, version: { type: integer, minimum: 1 }, attributes: { type: object, additionalProperties: true } }
...
    UuidV7:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
      description: >-
        Canonical lowercase UUID version 7 generated by PVLog. Treat this value
        as opaque; its time-sortable bits do not grant authority or establish an
        authoritative creation timestamp.
      examples:
        - 019505c8-7c85-7f0b-9bc3-2a3c4d5e6f70
...
    Problem:
      type: object
      additionalProperties: true
      description: RFC 9457 problem details.
      required:
        - type
        - title
        - status
        - requestId
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable problem type identifier.
        title:
          type: string
          description: Short human-readable problem title.
        status:
          type: integer
          minimum: 400
          maximum: 599
          description: HTTP status code.
        detail:
          type: string
          description: Safe occurrence-specific explanation.
        requestId:
          $ref: "#/components/schemas/UuidV7"